ref: Restructure exports to prefer default exports
