+++ packages/core/src/lib/error.ts
@@ -1,0 +1,9 @@
+export default class SentryError extends Error {
+  public name: string;
+
+  constructor(public message: string) {
+    super(message);
+    this.name = new.target.prototype.constructor.name;
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
