+++ packages/node/src/index.ts
@@ -1,3 +1,3 @@
-export { SentryNode, NodeOptions as SentryNodeOptions } from './lib/node';
+export { default, NodeOptions } from './lib/node';
 export { default as Store } from './lib/store';
 export { mkdirp, mkdirpSync } from './lib/utils';
